// frontend/clients/generated/instructions/deposit.ts
export type DepositInstructionAccounts = {
  owner: Signer;
  market: PublicKey | Pda;
  authority?: PublicKey | Pda;
  assetRegistry?: PublicKey | Pda;
  pool: PublicKey | Pda;
  ownerTokenAta: PublicKey | Pda;
  poolVault: PublicKey | Pda;
  vaultAuth?: PublicKey | Pda;
  obligation?: PublicKey | Pda;
  tokenProgram?: PublicKey | Pda;
  systemProgram?: PublicKey | Pda;
};

// Data.
export type DepositInstructionData = {
  discriminator: Uint8Array;
  amount: bigint;
};

export type DepositInstructionDataArgs = { amount: number | bigint };
// ...
export type DepositInstructionArgs = DepositInstructionDataArgs;